requirements.txt
# ============================================
# Python Dependencies
# Real-time Resource Usage Analysis
# ============================================

# --- Kafka ---
kafka-python==2.0.2

# --- Spark ---
pyspark==3.5.0

# --- ML / Data Science ---
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0

# --- Deep Learning ---
torch==2.2.0
torchvision==0.17.0

# --- Visualization ---
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# --- Dashboard ---
streamlit==1.31.0

# --- Utilities ---
tqdm==4.66.1
